+++ packages/dash-core-components/CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.28.2] - 2018-09-07
+## [0.28.3] - 2018-09-07
 ### Changed
 - The `Interval` component's `max_interval` prop can now be used to stop/restart the interval. Fixes [#266](https://github.com/plotly/dash-core-components/issues/266)
 - The `Graph` component's `id` is now not required to be set.